% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_c_hat}
\alias{get_c_hat}
\title{get_c_hat}
\usage{
get_c_hat(b1_hat, b2_hat, A, r)
}
\arguments{
\item{b1_hat}{estimated effect size of the first phenotype across all variants}

\item{b2_hat}{estimated effect size of the second phenotype across all variants}

\item{A}{Allele frequency information}

\item{r}{phenotypic correlation between the two phenotypes}
}
\value{
the MLE of slope between two sets of effect sizes
}
\description{
ALLSPICE function: compute the slope `c` that maximize the likelihood (maximum likelihood estimate - MLE)
}
