% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pheno_pair}
\alias{get_pheno_pair}
\title{get_pheno_pair}
\usage{
get_pheno_pair(b, X, r)
}
\arguments{
\item{b}{true effect size matrix of variants on the two phenotypes}

\item{X}{genotype matrix}

\item{r}{phenotypic correlation between the two phenotypes}
}
\value{
A 2x`n_ind` matrix of phenotype information (first row corresponds to the first phenotype, second row corresponds to the second phenotype)
}
\description{
simulation function: simulate true phenotype values of a pair of phenotypes
}
